requirements.txt
Flask==1.1.4
pyOpenSSL==20.0.1
python-telegram-bot==13.6
requests==2.25.1
urllib3==1.26.5
